--- src/event/mod.rs
@@ -1,7 +1,0 @@
-pub mod bus;
-
-pub mod base;
-pub mod strategy;
-
-pub use base::{Event, EventKind};
-pub use bus::Bus;
