+++ src/lib.rs
@@ -53,7 +53,7 @@
 //! │  {                                     │
 //! │    task: TaskRef,                      │
 //! │    restart: RestartPolicy,             │
-//! │    backoff: BackoffStrategy,           │
+//! │    backoff: BackoffPolicy,             │
 //! │    timeout: Option<Duration>           │
 //! │  }                                     │
 //! └────┬───────────────────────────────────┘
@@ -77,7 +77,7 @@
 //!      │                      - OnFailure    ⇒ retry
 //!      │                      - Always       ⇒ retry
 //!      │ (4) if retry:
-//!      │       delay = backoff.next(prev_delay) // BackoffStrategy from TaskSpec
+//!      │       delay = backoff.next(prev_delay) // BackoffPolicy from TaskSpec
 //!      │       publish BackoffScheduled{ task, delay, attempt, error }
 //!      │       sleep(delay)  (cancellable via runtime token)
 //!      │       prev_delay = Some(delay)
@@ -95,21 +95,21 @@
 //! | Area              | Description                                                           | Key types / traits                     |
 //! |-------------------|-----------------------------------------------------------------------|----------------------------------------|
 //! | **Observer API**  | Hook into task lifecycle events (logging, metrics, custom observers). | [`Observer`]                           |
-//! | **Policies**      | Configure restart/backoff strategies for tasks.                       | [`RestartPolicy`], [`BackoffStrategy`] |
+//! | **Policies**      | Configure restart/backoff strategies for tasks.                       | [`RestartPolicy`], [`BackoffPolicy`]   |
 //! | **Supervision**   | Manage groups of tasks and their lifecycle.                           | [`Supervisor`]                         |
 //! | **Errors**        | Typed errors for orchestration and task execution.                    | [`TaskError`], [`RuntimeError`]        |
 //! | **Tasks**         | Define tasks as functions or specs, easy to compose and run.          | [`TaskRef`], [`TaskFn`], [`TaskSpec`]  |
 //! | **Configuration** | Centralize runtime settings.                                          | [`Config`]                             |
 //!
 //! ## Optional features
-//! - `logging`: exports a simple built-in [`LoggerObserver`] _(demo/reference only)_.
+//! - `logging`: exports a simple built-in [`Writer`] _(demo/reference only)_.
 //! - `events`:  exports [`Event`] and [`EventKind`] for advanced integrations.
 //!
 //! ```no_run
 //! use std::time::Duration;
 //! use tokio_util::sync::CancellationToken;
 //! use taskvisor::{
-//!     BackoffStrategy, Config, RestartPolicy, Supervisor, TaskFn, TaskRef, TaskSpec, LoggerObserver
+//!     BackoffPolicy, Config, RestartPolicy, Supervisor, TaskFn, TaskRef, TaskSpec, LogWriter
 //! };
 //!
 //! #[tokio::main(flavor = "current_thread")]
@@ -118,7 +118,7 @@
 //!     cfg.timeout = Duration::from_secs(5);
 //!
 //!     // Use the built-in logger observer (enabled via --features "logging").
-//!     let s = Supervisor::new(cfg.clone(), LoggerObserver);
+//!     let s = Supervisor::new(cfg.clone(), LogWriter);
 //!
 //!     // Define a simple task with a cancellation token.
 //!     let hello: TaskRef = TaskFn::arc("hello", |ctx: CancellationToken| async move {
@@ -131,7 +131,7 @@
 //!     let spec = TaskSpec::new(
 //!         hello,
 //!         RestartPolicy::Never,
-//!         BackoffStrategy::default(),
+//!         BackoffPolicy::default(),
 //!         Some(Duration::from_secs(5)),
 //!     );
 //!
@@ -143,26 +143,26 @@
 mod config;
 mod core;
 mod error;
-mod event;
+mod events;
 mod observers;
 mod policy;
 mod task;
 // ---- Public re-exports ----
 
 pub use config::Config;
-pub use core::supervisor::Supervisor;
+pub use core::Supervisor;
 pub use error::{RuntimeError, TaskError};
-pub use event::strategy::BackoffStrategy;
-pub use observers::observer::Observer;
+pub use observers::Observer;
+pub use policy::BackoffPolicy;
 pub use policy::RestartPolicy;
 pub use task::{Task, TaskFn, TaskRef, TaskSpec};
 
 // Optional: expose event types.
 // Enable with: `--features events`
 #[cfg(feature = "events")]
-pub use crate::event::{Event, EventKind};
+pub use crate::events::{Event, EventKind};
 
 // Optional: expose a simple built-in logger observer (demo/reference).
 // Enable with: `--features logging`
 #[cfg(feature = "logging")]
-pub use observers::observer::LoggerObserver;
+pub use observers::LogWriter;
