+++ src/observers/mod.rs
@@ -1,2 +1,7 @@
-pub mod alive;
-pub mod observer;
+mod alive;
+mod log;
+mod observer;
+
+pub use alive::AliveTracker;
+pub use log::LogWriter;
+pub use observer::Observer;
