+++ src/task/spec.rs
@@ -7,7 +7,7 @@
 use std::fmt;
 use std::time::Duration;
 
-use crate::event::strategy::BackoffStrategy;
+use crate::policy::BackoffPolicy;
 use crate::task::base::TaskRef;
 use crate::{config::Config, policy::RestartPolicy};
 
@@ -16,7 +16,7 @@ use crate::{config::Config, policy::RestartPolicy};
 /// A [`TaskSpec`] bundles:
 /// - the task itself ([`TaskRef`])
 /// - restart policy ([`RestartPolicy`])
-/// - backoff strategy ([`BackoffStrategy`])
+/// - backoff polict ([`BackoffPolicy`])
 /// - optional execution timeout
 ///
 /// It can be created manually with [`TaskSpec::new`] or derived from a
@@ -25,7 +25,7 @@ use crate::{config::Config, policy::RestartPolicy};
 /// # Example
 /// ```
 /// use tokio_util::sync::CancellationToken;
-/// use taskvisor::{TaskSpec, TaskFn, Config, RestartPolicy, BackoffStrategy, TaskRef};
+/// use taskvisor::{TaskSpec, TaskFn, Config, RestartPolicy, BackoffPolicy, TaskRef};
 ///
 /// let demo: TaskRef = TaskFn::arc("demo", |_ctx: CancellationToken| async move { Ok(()) });
 ///
@@ -33,7 +33,7 @@ use crate::{config::Config, policy::RestartPolicy};
 /// let spec = TaskSpec::new(
 ///     demo.clone(),
 ///     RestartPolicy::Never,
-///     BackoffStrategy::default(),
+///     BackoffPolicy::default(),
 ///     None,
 /// );
 /// assert!(spec.timeout.is_none());
@@ -48,8 +48,8 @@ pub struct TaskSpec {
     pub task: TaskRef,
     /// Policy controlling if/when the task should be restarted.
     pub restart: RestartPolicy,
-    /// Strategy controlling delays between restarts.
-    pub backoff: BackoffStrategy,
+    /// Policy controlling delays between restarts.
+    pub backoff: BackoffPolicy,
     /// Optional timeout for the task execution.
     pub timeout: Option<Duration>,
 }
@@ -59,7 +59,7 @@ impl TaskSpec {
     pub fn new(
         task: TaskRef,
         restart: RestartPolicy,
-        backoff: BackoffStrategy,
+        backoff: BackoffPolicy,
         timeout: Option<Duration>,
     ) -> Self {
         Self {
